Implemented camera functions

+++ include/graphics/Camera.h
@@ -11,7 +11,7 @@ private:
 
 public:
     Camera(float aspectRatio);
-    void rotate(float x, float y, float z, float deg);
+    void rotate(int x, int y, int z, float deg);
     void move(float x, float y, float z);
     QMatrix4x4 getMatrix();
 };
